Only user who created a task can delete it

+++ app/Http/Controllers/TasksController.php
@@ -57,6 +57,8 @@ class TasksController extends Controller
     {
         $task = Task::findOrFail($id);
 
+        $this->authorize('deleteTask', $task);
+
         $task->delete();
 
         return redirect(route("home"));
